refactor index file and add tests
