+++ lib/retrieve_package.js
@@ -4,8 +4,6 @@ var request = require('request');
 var log = require('npmlog');
 var mkdirp = require('mkdirp');
 var existsAsync = fs.exists || path.exists;
-var d3 = require('d3-queue');
-
 
 function place_binary(options, callback) {
   var req = request.get({
@@ -72,31 +70,7 @@ function checkLibraryExists(options, callback) {
   });
 }
 
-function install(packageList, callback) {
-  var libraries = packageList;
-  var q = d3.queue(1);
-
-  libraries.forEach(function(options, i) {
-    options.update_binary = null;
-    if (options) {
-      checkLibraryExists(options, function(err, exists) {
-        if (!exists) {
-          log.info('check', 'checked for ' + options.name + ' (not found locally)');
-          q.defer(place_binary, options);
-          if (libraries.length - 1 === i) {
-            q.awaitAll(function(err) {
-              if (err) return callback(err);
-              return callback(null);
-            });
-          }
-        }
-      });
-    }
-  });
-}
-
 module.exports = {
   place_binary: place_binary,
-  install: install,
   checkLibraryExists: checkLibraryExists
 };
